DHCP agent: support routed VM interfaces as well as bridged
This patch shows a WIP set of changes to the reference Neutron DHCP
agent, for a network implementation in which data to and from VMs is
routed on the compute host instead of being bridged.  It is a logical
accompaniment to (although not strictly implied by) Carl Baldwin's
work on L3 routed networks at [1].

[1] https://review.openstack.org/#/c/196812/

Network implementation context
==============================

In the network implementation for which these DHCP agent changes are
designed:

- the host end of each TAP interface to/from a VM is left unbridged,
  and without an IPv4 address

- therefore, data _from_ a VM falls into the Linux network stack and
  is either delivered locally (if addressed to a local IP address) or
  routed

- the network implementation also installs local routes, one through
  each TAP interface, which allow data to be routed _to_ the locally
  hosted VMs.

DHCP service
============

The question then arises of how to provide DHCP service to those TAP
interfaces, and that is what is answered by this patch.  Its main
points are as follows.

- The (current) trigger for activating all of the following behaviour
  is setting the DHCP agent's interface_driver setting to
  RoutedInterfaceDriver.

- When RoutedInterfaceDriver is in use:

  - Everything is done in the default namespace.

  - The DHCP agent invokes Dnsmasq with modified options, notably
    '--bridge-interface=ns-XXX,tap* --interface=tap*', that tell
    Dnsmasq to listen on all the TAP interfaces and to treat those
    interfaces as aliases of the ns-XXX interface where the DHCP
    context and IP subnet are defined.

  - The ns-XXX interface (for the DHCP port) is created as a dummy
    interface.

  - The ns-XXX interface is given the subnet's gateway IP address,
    instead of a freshly allocated IP address from the subnet.  Care
    is taken at this point to _avoid_ a corresponding IP prefix route
    for the subnet, via the dummy device; as that would obviously go
    nowhere, and it causes a problem known as 'AS puddling' if that
    route is inadvertently exported via BGP to other compute
    hosts (and if there are IP hops between compute hosts).

Status and remaining work
=========================

My team has been running successfully for several months with versions
of these changes based on Icehouse, Juno and Kilo code.  So in that
sense, we know that the changes work.

However I very much doubt that the modified behaviour is being
correctly triggered.  Instead of being triggered by DHCP agent
_config_ (interface_driver=RoutedInterfaceDriver), my feeling is that
it would be better for the trigger to be some flag or property of the
network object coming from the Neutron server/DB.

In [1], Carl Baldwin has proposed a boolean 'l3_only' flag; and Kevin
Benton has suggested a provider network type instead.  There's also
the possibility of using a network flavor - although I'm not clear if
this would be the kind of use case for which flavors are really
intended.

Note also that the meaning needed for the DHCP agent changes here -
i.e. that the TAP interfaces are unbridged - may not be exactly the
same as the meaning of a network-level L3 flag/value - i.e. that you
can't assume there is L2 broadcast connectivity between all ports on
this network.  Multiple flags/triggers may be needed to describe these
nuances.

So, although I doubt that they are close to final form, I've posted
these changes now for the following reasons.

- To support Carl's work at [1], in the form of a concrete set of
  reference implementation changes that I'd like to make, but which
  depend on the API-level justification and model - for L3-only
  networks - that Carl is describing.

- To solicit input on what the correct trigger for the new behaviour
  should be.

- To show the detailed code changes, in case someone feels like
  commenting on those in detail for any other reason.

Please do let me know what you think!

Change-Id: I4dc5ae75f722d079a529ea1ae5f2011a3edbc84e